requirements.txt
@"
Django>=4.2
djangorestframework
django-cors-headers
Pillow
gunicorn
whitenoise
psycopg2-binary
python-dotenv
dj-database-url
"@ | Out-File -FilePath requirements.txt -Encoding UTF8